git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@8106 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/fix.h
@@ -148,7 +148,7 @@ class Fix : protected Pointers {
 
   virtual void read_data_header(char *) {}
   virtual void read_data_section(char *, int, char *) {}
-  virtual int read_data_skip_lines(char *) {}
+  virtual bigint read_data_skip_lines(char *) {}
 
   virtual int modify_param(int, char **) {return 0;}
